Use design system page title component Move version picker to misc
